This revision implements incomplete support for Visual Studio 2015.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3285 42547f41-83fc-0310-9aef-9668e019ab9e

+++ tools/win32/src/mxgnuplt.c
@@ -6,7 +6,7 @@
  *
  *---------------------------------------------------------------------------
  *
- * Copyright 1999, 2004, 2006, 2010 Illinois Institute of Technology
+ * Copyright 1999, 2004, 2006, 2010, 2015 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -181,7 +181,7 @@ main( int argc, char *argv[] )
 			if ( status != 0 ) {
 				sprintf( buffer,
 	"mxgnuplt: Error sending char '%c' to Gnuplot.  status = %ld\n",
-					c, status );
+					c, (long) status );
 				MessageBox( NULL, (LPTSTR) buffer,
 						NULL, MB_OK );
 			}
